NC 16 fix for webrtc.css
Quick 'n dirty fix for Speed.ME app UI. This enable to show the main sharing area in Nextcloud 16. Not perfect, makes you scrolling for chat input - in return it makes the "navbar" and "call bar" invisible - which necessary for picking up an incoming call or stop existing one. Until final fix: scrolling it is...